Rename queue to queue.v2
