+++ app/api/queue.v2/application/specs/Queue.spec.ts
@@ -1,7 +1,7 @@
 /* eslint-disable max-statements */
-import { Job } from 'api/queue/contracts/Job';
-import { MemoryQueueAdapter } from 'api/queue/infrastructure/MemoryQueueAdapter';
-import { StringJobSerializer } from 'api/queue/infrastructure/StringJobSerializer';
+import { Job } from 'api/queue.v2/contracts/Job';
+import { MemoryQueueAdapter } from 'api/queue.v2/infrastructure/MemoryQueueAdapter';
+import { StringJobSerializer } from 'api/queue.v2/infrastructure/StringJobSerializer';
 import { Queue } from '../Queue';
 
 async function sleep(ms: number) {
