+++ app/api/queue.v2/contracts/QueueAdapter.ts
