+++ app/worker.ts
@@ -12,9 +12,9 @@ import { syncWorker } from 'api/sync/syncWorker';
 import { InformationExtraction } from 'api/services/informationextraction/InformationExtraction';
 import { setupWorkerSockets } from 'api/socketio/setupSockets';
 import { ConvertToPdfWorker } from 'api/services/convertToPDF/ConvertToPdfWorker';
-import { QueueWorker } from 'api/queue/application/QueueWorker';
-import { Queue } from 'api/queue/application/Queue';
-import { StringJobSerializer } from 'api/queue/infrastructure/StringJobSerializer';
+import { QueueWorker } from 'api/queue.v2/application/QueueWorker';
+import { Queue } from 'api/queue.v2/application/Queue';
+import { StringJobSerializer } from 'api/queue.v2/infrastructure/StringJobSerializer';
 import Redis from 'redis';
 import RedisSMQ from 'rsmq';
 import { UpdateRelationshipPropertiesJob } from 'api/relationships.v2/services/propertyUpdateStrategies/UpdateRelationshipPropertiesJob';
